+++ src/main/kotlin/sundaystudy/kr/usedcar/module/bookmark/entity/Bookmark.kt
@@ -3,7 +3,8 @@ package sundaystudy.kr.usedcar.module.bookmark.entity
 import jakarta.persistence.*
 import sundaystudy.kr.usedcar.global.audit.Auditable
 import sundaystudy.kr.usedcar.global.audit.BaseTime
-import sundaystudy.kr.usedcar.module.post.entity.Post
+import sundaystudy.kr.usedcar.module.usedcar.entity.Post
+
 import sundaystudy.kr.usedcar.module.user.entity.User
 
 @Entity
