--- src/main/kotlin/sundaystudy/kr/usedcar/module/insurance/entity/ChangeOwner.kt
@@ -1,20 +1,0 @@
-package sundaystudy.kr.usedcar.module.insurance.entity
-
-import jakarta.persistence.*
-import java.time.LocalDateTime
-import java.util.*
-
-@Entity
-class ChangeOwner(
-    @Column(name = "change_day")
-    var changeDay: LocalDateTime,
-) {
-
-    @Id
-    @Column(columnDefinition = "BINARY(16)")
-    var id: UUID = UUID.randomUUID()
-
-    @JoinColumn(name = "insuracne_id")
-    @ManyToOne(fetch = FetchType.LAZY)
-    var insurance: Insurance? = null
-}
