--- src/main/kotlin/sundaystudy/kr/usedcar/module/insurance/repository/UnsubscribedRepository.kt
@@ -1,7 +1,0 @@
-package sundaystudy.kr.usedcar.module.insurance.repository
-
-import org.springframework.data.jpa.repository.JpaRepository
-import sundaystudy.kr.usedcar.module.insurance.entity.Unsubscribed
-import java.util.UUID
-
-interface UnsubscribedRepository : JpaRepository<Unsubscribed, UUID>
