+++ src/main/kotlin/sundaystudy/kr/usedcar/module/matching/entity/Matching.kt
@@ -3,7 +3,7 @@ package sundaystudy.kr.usedcar.module.matching.entity
 import jakarta.persistence.*
 import sundaystudy.kr.usedcar.global.audit.Auditable
 import sundaystudy.kr.usedcar.global.audit.BaseTime
-import sundaystudy.kr.usedcar.module.post.entity.Post
+import sundaystudy.kr.usedcar.module.usedcar.entity.Post
 import sundaystudy.kr.usedcar.module.user.entity.User
 
 @Entity
