--- src/main/kotlin/sundaystudy/kr/usedcar/module/post/repository/CarRepository.kt
@@ -1,7 +1,0 @@
-package sundaystudy.kr.usedcar.module.post.repository
-
-import org.springframework.data.jpa.repository.JpaRepository
-import sundaystudy.kr.usedcar.module.post.entity.Car
-import java.util.UUID
-
-interface CarRepository : JpaRepository<Car, UUID>
