--- src/main/kotlin/sundaystudy/kr/usedcar/module/post/repository/UsedCarRepository.kt
@@ -1,7 +1,0 @@
-package sundaystudy.kr.usedcar.module.post.repository
-
-import org.springframework.data.jpa.repository.JpaRepository
-import sundaystudy.kr.usedcar.module.post.entity.UsedCar
-import java.util.UUID
-
-interface UsedCarRepository : JpaRepository<UsedCar, UUID>
