+++ src/main/kotlin/sundaystudy/kr/usedcar/module/usedcar/dto/response/UsedCarResponse.kt
@@ -1,0 +1,10 @@
+import sundaystudy.kr.usedcar.module.usedcar.dto.response.CarResponse
+import sundaystudy.kr.usedcar.module.usedcar.insurance.dto.response.InsuranceResponse
+
+data class UsedCarResponse(
+    var licenseNumber: String,
+    var price: Int,
+    var savePrice: Int,
+    var carResponse: CarResponse,
+    var insuranceDetail : InsuranceResponse
+)
