+++ src/main/kotlin/sundaystudy/kr/usedcar/module/usedcar/insurance/dto/detail/ChangeNumberDetail.kt
@@ -1,0 +1,9 @@
+package sundaystudy.kr.usedcar.module.usedcar.insurance.dto.detail
+
+import java.time.LocalDateTime
+
+data class ChangeNumberDetail(
+    var changeDay: LocalDateTime,
+    var changeName: String,
+    var isFirst: Boolean
+)
