+++ src/main/kotlin/sundaystudy/kr/usedcar/module/usedcar/insurance/dto/detail/ChangeOwnerDetail.kt
@@ -1,0 +1,7 @@
+package sundaystudy.kr.usedcar.module.usedcar.insurance.dto.detail
+
+import java.time.LocalDateTime
+
+data class ChangeOwnerDetail(
+    var changeDay: LocalDateTime
+)
