+++ src/main/kotlin/sundaystudy/kr/usedcar/module/usedcar/insurance/dto/detail/UnsubscribedDetail.kt
@@ -1,0 +1,8 @@
+package sundaystudy.kr.usedcar.module.usedcar.insurance.dto.detail
+
+import java.util.*
+
+data class UnsubscribedDetail(
+    var startAt: Date,
+    var endAt: Date
+)
