+++ src/main/kotlin/sundaystudy/kr/usedcar/module/usedcar/insurance/entity/ChangeOwner.kt
@@ -1,0 +1,20 @@
+package sundaystudy.kr.usedcar.module.usedcar.insurance.entity
+
+import jakarta.persistence.*
+import java.time.LocalDateTime
+import java.util.*
+
+@Entity
+class ChangeOwner(
+    @Column(name = "change_day")
+    var changeDay: LocalDateTime,
+) {
+
+    @Id
+    @Column(columnDefinition = "BINARY(16)")
+    var id: UUID = UUID.randomUUID()
+
+    @JoinColumn(name = "insuracne_id")
+    @ManyToOne(fetch = FetchType.LAZY)
+    var insurance: Insurance? = null
+}
