+++ src/main/kotlin/sundaystudy/kr/usedcar/module/usedcar/repository/UsedCarRepository.kt
@@ -1,0 +1,7 @@
+package sundaystudy.kr.usedcar.module.usedcar.repository
+
+import org.springframework.data.jpa.repository.JpaRepository
+import sundaystudy.kr.usedcar.module.usedcar.entity.UsedCar
+import java.util.UUID
+
+interface UsedCarRepository : JpaRepository<UsedCar, UUID>
